refactor(hooks): lift the picks fetch out of usePlayerScores
loadStoredPicks touched no state, props, or refs, so the useCallback around it
was ceremony. Out of the hook it is testable without rendering.
