+++ src/utils/CLAUDE.md
@@ -5,6 +5,7 @@
 - `buildSpreadsheetBuffer`: the xlsx-js-style workbook export and its content type
 - `pickStatusFill`: pick colors for the export
 - `picksCache`: localStorage cache of an uploaded workbook, per season and week
+- `loadStoredPicks`: a week's workbook from the API, falling back to that cache
 - `debugLog`: scoring traces, silent outside a dev server
 - `getClasses`: conditional className join
 - `plural`: a count and its noun, pluralized
